feat: Add paid_user field to User model
This commit adds the `paid_user` field to the `User` model in both the server and web packages. The `paid_user` field is a boolean value that indicates whether a user is a paid user or not. This change allows for distinguishing between free and paid users in the application.